refactor: dedupe test_new_all_broadcast helpers into shared module (-929/+30 lines across 4 copies)
Move run_with_trace/run_all_broadcast_impl from the four copies (box/all_post_commit,
box/nightly, galaxy/galaxy_nightly, galaxy/nightly) into
blackhole_CI/_all_broadcast_helpers.py. Drift between copies is preserved via
cache_counter_mode and placement_mode parameters; each file binds its
pre-refactor behavior via functools.partial. Tails unchanged.
