+++ tests/ttnn/unit_tests/operations/ccl/blackhole_CI/galaxy/galaxy_nightly/test_new_all_broadcast.py
@@ -2,242 +2,18 @@
 
 # SPDX-License-Identifier: Apache-2.0
 
-
-import torch
 import pytest
-from loguru import logger
 import ttnn
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_equal, comp_pcc
 from models.common.utility_functions import (
     skip_for_wormhole_b0,
     skip_for_n_or_less_dev,
 )
 
-
-def run_with_trace(
-    mesh_device,
-    all_broadcast_topology,
-    input_tensor_mesh,
-    num_links,
-    output_mem_config,
-    num_iter=20,
-    subdevice_id=None,
-    cluster_axis=None,
-):
-    # Compile Run
-    logger.info("Compiling model")
-    tt_out_tensor = ttnn.all_broadcast(
-        input_tensor_mesh,
-        num_links=num_links,
-        memory_config=output_mem_config,
-        topology=all_broadcast_topology,
-        subdevice_id=subdevice_id,
-        cluster_axis=cluster_axis,
-    )
-    ttnn.synchronize_device(mesh_device)
-
-    # Capture trace
-    logger.info("Capturing trace")
-    trace_id = ttnn.begin_trace_capture(mesh_device, cq_id=0)
-    for i in range(num_iter):
-        tt_out_tensor = ttnn.all_broadcast(
-            input_tensor_mesh,
-            num_links=num_links,
-            memory_config=output_mem_config,
-            topology=all_broadcast_topology,
-            subdevice_id=subdevice_id,
-            cluster_axis=cluster_axis,
-        )
-    ttnn.end_trace_capture(mesh_device, trace_id, cq_id=0)
-    ttnn.synchronize_device(mesh_device)
-
-    # Run the op
-    logger.info("Starting Trace perf test...")
-    ttnn.execute_trace(mesh_device, trace_id, blocking=False)
-    ttnn.release_trace(mesh_device, trace_id)
-    ttnn.synchronize_device(mesh_device)
-
-    return tt_out_tensor
+# The shared helper defaults (no cache-entries counter, fixed 8x4 mesh-mapper
+# placement) match this file's pre-refactor behavior.
+from tests.ttnn.unit_tests.operations.ccl.blackhole_CI._all_broadcast_helpers import run_all_broadcast_impl
 
 
-def run_all_broadcast_impl(
-    mesh_device,
-    num_devices,
-    output_shape,
-    num_links,
-    input_dtype,
-    layout,
-    function_level_defaults,
-    all_broadcast_topology,
-    num_iters=1,
-    trace_mode=False,
-    rand_tensor=True,
-    mem_config=None,
-    input_shard_shape=None,
-    input_shard_grid=None,
-    output_shard_shape=None,
-    output_shard_grid=None,
-    tensor_mem_layout=None,
-    cluster_axis=None,
-):
-    if num_iters < 1:
-        pytest.fail("num_iters must be >= 1")
-
-    compute_grid_size = mesh_device.compute_with_storage_grid_size()
-    ccl_sub_device_crs = ttnn.CoreRangeSet(
-        {ttnn.CoreRange(ttnn.CoreCoord(0, 0), ttnn.CoreCoord(compute_grid_size.x - 1, compute_grid_size.y - 1))}
-    )
-    worker_sub_device = ttnn.SubDevice(
-        [
-            ccl_sub_device_crs,
-        ]
-    )
-    worker_sub_device_id = ttnn.SubDeviceId(0)
-    sub_device_stall_group = [worker_sub_device_id]
-    sub_device_manager = mesh_device.create_sub_device_manager([worker_sub_device], 0)
-    mesh_device.load_sub_device_manager(sub_device_manager)
-    mesh_device.set_sub_device_stall_group(sub_device_stall_group)
-
-    logger.info(f"Output shape: {output_shape}")
-    logger.info(f"input_shard_shape: {input_shard_shape}")
-    logger.info(f"input_shard_grid: {input_shard_grid}")
-
-    ### For sharded all broadcast only
-    if bool(input_shard_shape) != bool(input_shard_grid) and bool(tensor_mem_layout) != bool(input_shard_grid):
-        pytest.fail(
-            "Both input_shard_shape, shard_grid, and tensor_mem_layout must be provided together or all must be None"
-        )
-    if input_shard_shape and input_shard_grid:
-        input_shard_spec = ttnn.ShardSpec(
-            input_shard_grid,
-            input_shard_shape,
-            ttnn.ShardOrientation.ROW_MAJOR,
-        )
-        input_mem_config = ttnn.MemoryConfig(
-            tensor_mem_layout, buffer_type=ttnn.BufferType.L1, shard_spec=input_shard_spec
-        )
-        if output_shard_shape is None:
-            assert (
-                output_shard_grid is None
-            ), "output_shard_grid must not be provided if output_shard_shape is not provided"
-            output_shard_shape = input_shard_shape
-            output_shard_spec = ttnn.ShardSpec(
-                input_shard_grid,
-                output_shard_shape,
-                ttnn.ShardOrientation.ROW_MAJOR,
-            )
-            output_mem_config = ttnn.MemoryConfig(
-                tensor_mem_layout, buffer_type=ttnn.BufferType.L1, shard_spec=output_shard_spec
-            )
-        else:
-            assert output_shard_grid is not None, "output_shard_grid must be provided if output_shard_shape is provided"
-            output_shard_spec = ttnn.ShardSpec(
-                output_shard_grid,
-                output_shard_shape,
-                ttnn.ShardOrientation.ROW_MAJOR,
-            )
-            output_mem_config = ttnn.MemoryConfig(
-                tensor_mem_layout, buffer_type=ttnn.BufferType.L1, shard_spec=output_shard_spec
-            )
-    else:
-        assert mem_config is not None
-        input_mem_config = mem_config
-        output_mem_config = mem_config
-    ###
-
-    input_tensor_mesh_list = []
-    output_tensor_goldens_list = []
-
-    for i in range(num_iters):
-        output_tensors = []
-        for k in range(num_devices):
-            if rand_tensor:
-                output_tensor = torch.rand(output_shape).bfloat16()
-            else:
-                output_tensor = torch.zeros(output_shape)
-                row_id = 1
-                # Fix indices for tensors with ranks 2 or 3
-                for w in range(output_shape[0]):
-                    for z in range(output_shape[1]):
-                        for y in range(0, output_shape[2], 32):
-                            for x in range(0, output_shape[3], 32):
-                                output_tensor[w, z, y, :] = row_id
-                                row_id += 1
-            output_tensors.append(output_tensor)
-
-        output_tensor_goldens_list.append(output_tensors)
-        temp_output_tensor = torch.cat(output_tensors, -1)
-
-        input_tensor_mesh = ttnn.from_torch(
-            temp_output_tensor,
-            device=mesh_device,
-            layout=layout,
-            dtype=input_dtype,
-            memory_config=input_mem_config,
-            mesh_mapper=ttnn.create_mesh_mapper(
-                mesh_device,
-                ttnn.MeshMapperConfig([ttnn.PlacementReplicate(), ttnn.PlacementShard(-1)], ttnn.MeshShape(8, 4)),
-            ),
-        )
-
-        input_tensor_mesh_list.append(input_tensor_mesh)
-
-    tt_out_tensor_list = []
-    if trace_mode:
-        tt_out_tensor = run_with_trace(
-            mesh_device,
-            all_broadcast_topology,
-            input_tensor_mesh_list[0],
-            num_links,
-            output_mem_config,
-            num_iter=num_iters,
-            subdevice_id=worker_sub_device_id,
-            cluster_axis=cluster_axis,
-        )
-        tt_out_tensor_list.append(tt_out_tensor)
-    else:
-        for i in range(num_iters):
-            tt_out_tensors = ttnn.all_broadcast(
-                input_tensor_mesh_list[i],
-                num_links=num_links,
-                memory_config=output_mem_config,
-                topology=all_broadcast_topology,
-                subdevice_id=worker_sub_device_id,
-                cluster_axis=cluster_axis,
-            )
-            tt_out_tensor_list.append(tt_out_tensors)
-
-        logger.info(f"Waiting for op")
-        ttnn.synchronize_device(mesh_device, sub_device_ids=sub_device_stall_group)
-        logger.info(f"Done op")
-
-    passed = True
-    for tensor_index in range(len(tt_out_tensor_list)):
-        tt_out_tensors = tt_out_tensor_list[tensor_index]
-        output_tensors = output_tensor_goldens_list[tensor_index]
-        for k in range(num_devices):
-            output_tensor = output_tensors[k]
-            for i, t in enumerate(ttnn.get_device_tensors(tt_out_tensors[k])):
-                tt_output_tensor = t.cpu().to(ttnn.ROW_MAJOR_LAYOUT).to_torch()
-                logger.info(f"Checking for device {t.device().id()}")
-                if input_dtype == ttnn.bfloat16:
-                    eq, output = comp_equal(tt_output_tensor, output_tensor)
-                else:
-                    eq, output = comp_pcc(tt_output_tensor, output_tensor)
-                if not eq:
-                    logger.error(f"output mismatch for tensor {i}")
-                    passed = False
-                    assert eq, f"{i} FAILED: {output}"
-    assert (
-        mesh_device.num_program_cache_entries() == 1 or mesh_device.num_program_cache_entries() == num_iters
-    ), f"Device has {mesh_device.num_program_cache_entries()} program cache entries"
-    mesh_device.reset_sub_device_stall_group()
-    mesh_device.clear_loaded_sub_device_manager()
-    if not passed:
-        assert eq, f"{i} FAILED: {output}"
-
-
-# Enumerate the post-commit cases explicitly
 @skip_for_wormhole_b0()
 @skip_for_n_or_less_dev(1)
 @pytest.mark.parametrize(
